# public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/locations": {
      "post": {
        "operationId": "LocationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Locations"
        ]
      },
      "get": {
        "operationId": "LocationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locations"
        ]
      }
    },
    "/locations/{id}": {
      "get": {
        "operationId": "LocationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locations"
        ]
      },
      "patch": {
        "operationId": "LocationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locations"
        ]
      },
      "delete": {
        "operationId": "LocationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locations"
        ]
      }
    },
    "/agents": {
      "post": {
        "operationId": "AgentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Agents"
        ]
      },
      "get": {
        "operationId": "AgentsController_findAll",
        "parameters": [
          {
            "name": "location",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Agents"
        ]
      }
    },
    "/agents/{id}": {
      "get": {
        "operationId": "AgentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Agents"
        ]
      },
      "patch": {
        "operationId": "AgentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Agents"
        ]
      },
      "delete": {
        "operationId": "AgentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Agents"
        ]
      }
    }
  },
  "info": {
    "title": "Pickup Mtaani Demo Api",
    "description": "A demo api to demonstrate to pickup mtaani the api and endpoints that are required",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateLocationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "default": "Kisii",
            "description": "The name of the location. Can have one or more agents"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateLocationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "default": "Kisii",
            "description": "The name of the location. Can have one or more agents"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateAgentDto": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string"
          },
          "agent_name": {
            "type": "string"
          }
        },
        "required": [
          "location",
          "agent_name"
        ]
      },
      "UpdateAgentDto": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string"
          },
          "agent_name": {
            "type": "string"
          }
        },
        "required": [
          "location",
          "agent_name"
        ]
      }
    }
  }
}